requirements.txt
asgiref==3.2.3
Django==3.0.2
django-phone-field==1.8.0
pytz==2019.3
sqlparse==0.3.0
